services/CogMaker/requirements.txt
Flask
gunicorn
google-events
cloudevents
google-cloud-storage